chore(core): Pubsub parses incoming messages
