+++ packages/core/src/pubsub/task-queue.ts
@@ -21,7 +21,6 @@ export class TaskQueue {
    * Add task to queue.
    */
   add(f: Task<void>): void {
-    console.log('tq.add')
     this.#pq.add(f).catch((error) => {
       const retry = () => this.add(f);
       this.onError(error, retry);
@@ -32,7 +31,6 @@ export class TaskQueue {
    * Wait till all the tasks are completed.
    */
   onIdle(): Promise<void> {
-    console.log('tq.onIdle')
     return this.#pq.onIdle();
   }
